feat: Add method to register storage providers
Implementation

Add the `register_storage_provider` method to the storage provider
pallet. This method allows storage providers to register their
information on-chain. This method emits the `StorageProviderRegistered`
event which lists the static storage provider information and the owner
(`AccountId`) that the new storage provider belongs to. This method
fails if there is already a storage provider registered under the
incoming `AccountId`.

Storage provider information consists of 2 parts:

1. Static information
This is information about the storage provider that rarely changes, like
the peer_id and the PoSt proof type. This information is stored in the
`StorageProviderInfo` data structure.

2. Dynamic information
This is data that changes regularly, things like information about
sectors and pre commit deposits. This information is stored in the
`StorageProviderState` data structure. This data structure holds the
static information about the storage provider too.

All this information about the storage provider is stored in the
`StorageMap` called `StorageProviders`.

The pallets `Config` trait includes the following types:
- RuntimeEvent: This type is needed because the storage provider pallet
  emits events.
- PeerId: This type is the libp2p hash of the storage providers encoded
  public key, this is provided by the storage provider.
- Currency: This is the currency mechanism, used to store collateral for
  the pre committed sectors.

Tests

A mock runtime is constructed for running the tests. The test check if
the storage provider registration works and if it fails if the same
`AccountId` tries to register another storage provider.