test: Add tarpaulin to just and nix
The cargo-tarpaulin package has been added the `flake.nix` file.
The just commands have been updated to use `cargo tarpaulin` and produce an lcov and an HTML file.
The settings for the coverage gutters extension have been updated to reflect these changes.
An extra command has been added to the Justfile, `full-coverage` to run all the code coverage operations in a single command.