feat(market): Add storage provider validation
Add storage provider validation by adding the `StorageProvider` trait
and implementing it to the storage provider pallet so we can loosely
couple the market and storage provider pallet and call the
`is_registered_storage_provider` method in the market pallet.
The tests in the market pallet have been updated to include registering
a storage provider. A new test has been added to check the storage
provider registration.