fix(mater): ensure block stream returns properly sized blocks
Using `tokio::io::ReaderStream` yields a similar effect *iff*
the underlying stream is "reliable" — or in other words, is able
to fill the `ReaderStream` capacity in a single shot.

This is not always the case and lead to a really frustrating bug
when reading HTTP bodies, where the stream would return a smaller
buffer and the `ReaderStream` would hapilly pass it along.

I guess it was my bad since the `ReaderStream` documentation never
guaranteed that the chunks would have the "proper size".